requirements.txt
frappe
unidecode
pygithub